\name{bear}
\alias{bear}
\docType{data}
\title{GPS monitoring of one brown bear}
\description{
  These data contain the relocations of one female brown bear monitored
  using GPS collars during May 2004 in Sweden.
}
\usage{data(bear)}
\source{
Scandinavian Bear Research Project. Skandinaviska
Bjornprojektet. Tackasen - Kareliusvag 2. 79498 Orsa, Sweden.
email: info@bearproject.info
}
\examples{
data(bear)
plot(bear)
}
\keyword{datasets}
