\name{ibex}
\alias{ibex}
\docType{data}
\title{GPS Monitoring of Four Ibex in the Belledonne Mountain}
\description{
  This dataset is an object of class "ltraj" (regular trajectory,
  relocations every 4 hours) containing the GPS relocations of four ibex
  during 15 days in the Belledonne mountain (French Alps).
}
\usage{data(ibex)}
\source{
  Office national de la chasse et de la faune sauvage, CNERA Faune de
  Montagne, 95 rue Pierre Flourens, 34000 Montpellier, France.
}
\examples{
data(ibex)
plot(ibex)
}
\keyword{datasets}
