\name{mouflon}
\alias{mouflon}
\docType{data}
\title{GPS Monitoring of One Mouflon in the Caroux Mountain}
\description{
  This dataset is an object of class "ltraj" (regular trajectory,
  relocations every 20 minutes) containing the GPS relocations of one
  mouflon during two week-ends in the Caroux mountain (South of
  France).
}
\usage{data(mouflon)}
\source{
  Office national de la chasse et de la faune sauvage, CNERA Faune de
  Montagne, 95 rue Pierre Flourens, 34000 Montpellier, France.
}
\examples{
data(mouflon)
plot(mouflon)
}
\keyword{datasets}
