\name{puechcirc}
\alias{puechcirc}
\docType{data}
\title{Movements of wild boars tracked at Puechabon}
\description{
This data set is an object of class \code{ltraj}, giving the results of the
monitoring of 2 wild boars by radio-tracking at Puechabon (Mediterranean
habitat, South of France).  These data have been
collected by Daniel Maillard (Office national de la chasse et de la
faune sauvage), and correspond to the activity period of the wild boar
(during the night, when the animals forage. The data set
\code{puechabonsp} in the package \code{adehabitatMA} describes the
resting sites). 
}
\usage{data(puechcirc)}
\format{
  This object has, in total, 204 relocations distributed among two
  animals and three bursts of relocations (\code{CH930803}, \code{CH930824},
  \code{CH930827}, and \code{JE930827}).
}
\source{
Maillard, D. (1996). \emph{Occupation et utilisation de la garrigue et
  du vignoble mediterraneens par le Sanglier}.
Universite d'Aix-Marseille III: PhD thesis.
}
\keyword{datasets}
