\name{rupicabau}
\alias{rupicabau}
\docType{data}
\title{GPS Monitoring of One Chamois in the Bauges Mountains}
\description{
  This dataset is an object of class "ltraj" (regular trajectory,
  relocations every 20 minutes) containing the GPS relocations of two
  chamois during one day in the Bauges mountain (French Alps).
}
\usage{data(rupicabau)}
\source{
  Office national de la chasse et de la faune sauvage, CNERA Faune de
  Montagne, 95 rue Pierre Flourens, 34000 Montpellier, France.
}
\examples{
data(rupicabau)
plot(rupicabau)
}
\keyword{datasets}
